Uploading Films Web Application (First time)
This is films repository where initially a new user can be registered and can login. The dashboard contains all the films in the application with a pagination of 1 film per 1 page. The film has all the details in a tabular format including it's name, the photo, description, it's ticketing price, the rating according to certain criteria (1 to 5), the genre whether it's an action film or Sci-Fi, etc, and the release date. Once the user logs in, only then he/she can create a new film and can add a new comment to the films in the DB. Proper validations have been taken care everywhere while registering a new user, creating a new film and adding a new comment.